feat(trivia): automated daily question pre-generation via cron endpoint
- Extracts shared generation helpers (fetchOpenTDBQuestions, generateAIQuestion,
  generateFallbackQuestions, generateQuestionsForDate, daysSinceEpoch) into
  src/lib/trivia/generateDailyQuestions.ts
- Refactors daily/route.ts to import from the shared lib (no behavior change)
- Adds POST /api/v1/trivia/daily/generate?lookahead=N cron endpoint, protected
  by Bearer token (DAILY_TRIVIA_CRON_SECRET env var), skips already-generated
  dates idempotently

Closes #200
